optimize: Transaction timeout judgment is performed on the TM side and TC side

+++ tm/src/main/java/io/seata/tm/api/transaction/TransactionInfo.java
@@ -37,7 +37,7 @@ public final class TransactionInfo implements Serializable {
 
     private int lockRetryTimes;
 
-    private int lossTime;
+    private long lossTime;
 
     public int getTimeOut() {
         return timeOut;
@@ -110,11 +110,11 @@ public final class TransactionInfo implements Serializable {
         this.lockRetryTimes = lockRetryTimes;
     }
 
-    public int getLossTime() {
+    public long getLossTime() {
         return lossTime;
     }
 
-    public void setLossTime(int lossTime) {
+    public void setLossTime(long lossTime) {
         this.lossTime = lossTime;
     }
 }
